Updated files

+++ src/BmiService.java
@@ -1,8 +1,8 @@
 public class BmiService {
-    public int calculate (int weight, double height) {
+    public double calculate (int weight, double height) {
 
         double bmi = weight / height / height;
 
-        return (int) bmi;
+        return bmi;
     }
 }
